+++ src/parse.js
@@ -71,4 +71,17 @@ function getConfig() {
   return config;
 }
 
-module.exports = { parse };
+function validateVersion(input) {
+  // check that the version is valid
+  if (!/^\d+\.\d+\.\d+$/.test(input)) {
+    console.log(
+      "Invalid version: " +
+        input +
+        ". Please use the format MAJOR.MINOR.PATCH (e.g. 3.2.3)."
+    );
+    return null;
+  }
+  return input;
+}
+
+module.exports = { parse, validateVersion };
