--- src/utils/validate_version.js
@@ -1,14 +1,0 @@
-function validateVersion(input) {
-  // check that the version is valid
-  if (!/^\d+\.\d+\.\d+$/.test(input)) {
-    console.log(
-      "Invalid version: " +
-        input +
-        ". Please use the format MAJOR.MINOR.PATCH (e.g. 3.2.3)."
-    );
-    return null;
-  }
-  return input;
-}
-
-module.exports = validateVersion;
